car_showroom_project/requirements.txt
Django~=4.0.6
psycopg2~=2.9.3
djangorestframework~=3.13.1
django-countries~=7.3.2
django-filter~=22.1
django-debug-toolbar~=3.5.0
drf-yasg~=1.21.3
pytest~=7.1.2
pytest-django~=4.5.2
djangorestframework-simplejwt~=5.2.0